warning: comparison between signed and unsigned integer expressions fixed
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@2861 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ cdc/modules/cdcDigitizer/src/CDCDigiModule.cc
@@ -250,8 +250,8 @@ void CDCDigiModule::event()
     cdcSimHitToHitRel.add(iterCDCMap->second->getHitNumber(), iDigits);     //add entry
 
     for (int index = 0; index < cdcSimHitRel.getEntries(); index++) {
-      for (int hit = 0; hit < cdcSimHitRel[index].getToIndices().size(); hit++) {
-        if (cdcSimHitRel[index].getToIndex(hit) == (unsigned int) iterCDCMap->second->getHitNumber()) {
+      for (int hit = 0; hit < (int)cdcSimHitRel[index].getToIndices().size(); hit++) {
+        if ((int)cdcSimHitRel[index].getToIndex(hit) == iterCDCMap->second->getHitNumber()) {
           mcPartToCDCHitRel.add(cdcSimHitRel[index].getFromIndex(), iDigits);      //add entry
         }
       }
